[MOD] fragment&viewModel 코드 전환
