+++ app/src/main/java/com/puzzling/puzzlingaos/presentation/team/currentSituation/ItemTeamRetrospectListFragment.kt
@@ -1,15 +1,11 @@
 package com.puzzling.puzzlingaos.presentation.team.currentSituation
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModelProvider
 import com.puzzling.puzzlingaos.R
 import com.puzzling.puzzlingaos.base.BaseFragment
-import com.puzzling.puzzlingaos.data.model.response.ResponseTeamReviewListDto
 import com.puzzling.puzzlingaos.databinding.FragmentItemTeamRetrospectListBinding
-import com.puzzling.puzzlingaos.domain.entity.TeamReviewMultiList
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -33,14 +29,7 @@ class ItemTeamRetrospectListFragment(
         if (retroItem != null) {
             for (item in retroItem) {
                 if (day == item.reviewDay) {
-                    testItemList(viewModel.teamRetrospectList, day)
-                }
-            }
-        }
-
-        if (retroItem != null) {
-            for (item in retroItem) {
-                if (day == item.reviewDay) {
+                    viewModel.testItemList(viewModel.teamRetrospectList, day)
                     binding.rcvTeamRetrospectList.adapter = viewModel.teamRetrospectMultiList.value?.let {
                         RetrospectListAdapter(
                             it,
@@ -49,60 +38,5 @@ class ItemTeamRetrospectListFragment(
                 }
             }
         }
-    }
-
-    fun testItemList(itemList: LiveData<ArrayList<ResponseTeamReviewListDto.Data>>, day: String) {
-        var intDay: Int = viewModel.itemRetroList.indexOfFirst { it.reviewDay == day }
-
-        viewModel.teamRetrospectMultiList.value?.clear()
-
-        if (itemList.value?.isNotEmpty() == true) {
-            if (itemList.value?.get(intDay)?.reviewWriters?.isNotEmpty() == true) {
-                viewModel.teamRetrospectMultiList.value?.add(
-                    TeamReviewMultiList(
-                        0,
-                        null,
-                        null,
-                        null,
-                        null,
-                    ),
-                )
-                for (i in 0 until (itemList.value?.get(intDay)?.reviewWriters!!.size ?: 0)) {
-                    viewModel.teamRetrospectMultiList.value?.add(
-                        TeamReviewMultiList(
-                            1,
-                            itemList.value?.get(intDay)?.reviewDay,
-                            itemList.value?.get(intDay)?.reviewDate,
-                            itemList.value?.get(intDay)?.reviewWriters?.get(i)?.memberNickname,
-                            itemList.value?.get(intDay)?.reviewWriters?.get(i)?.memberRole,
-                        ),
-                    )
-                }
-            }
-
-            if (itemList.value?.get(intDay)?.nonReviewWriters?.isNotEmpty() == true) {
-                viewModel.teamRetrospectMultiList.value?.add(
-                    TeamReviewMultiList(
-                        2,
-                        null,
-                        null,
-                        null,
-                        null,
-                    ),
-                )
-                for (i: Int in 0 until (itemList.value?.get(intDay)?.nonReviewWriters?.size ?: 0)) {
-                    viewModel.teamRetrospectMultiList.value?.add(
-                        TeamReviewMultiList(
-                            3,
-                            itemList.value?.get(intDay)?.reviewDay,
-                            itemList.value?.get(intDay)?.reviewDate,
-                            itemList.value?.get(intDay)?.nonReviewWriters?.get(i)?.memberNickname,
-                            itemList.value?.get(intDay)?.nonReviewWriters?.get(i)?.memberRole,
-                        ),
-                    )
-                }
-            }
-        }
-        Log.d("ItemTeamRetrospectListFragment: ", "${viewModel.teamRetrospectMultiList.value}")
     }
 }
